Move neighbor module to _neighbor
