+++ signac/_neighbor.py
